federated_platform_v2/requirements.txt
grpcio
grpcio-tools
protobuf
numpy
torch
opacus
pyyaml
scikit-learn
prometheus-client
tenacity
hydra-core
